Refactored tests into different files for easier execution
