+++ OdinAPI/tests/newUserData.py
@@ -1,0 +1,8 @@
+newUser = {
+        'username': 'newUser17',
+        'email': 'newuser17@email.com',
+        'full_name': 'New User17',
+        'password':'someR4nd0mP455woRd',
+      }
+
+newUserID = 49
